Added itsLearning-Login to `index5.js`
This overwrites the default login, which uses the own DB, and instead
uses the itsLearning-API to verify the user-credentials.

The main problem that arises through this configuration is that the API
seems to not pass an equivalent to "user_class", thus 'destroying'
almost every part of Schoolist, as it relies on "user_class"!